Added crude "water tracking" capability to DVL plugin
Selectable via the plugin definition <enableWaterTrack> element in
the plugin definition (default to disabled).  The sensor keeps track
of both water track and bottom track velocities.  Water track velocity
is currently based on linear velocity and does not account for any
sort of current (possible future work).  When enabled, the plugin
will use bottom track if it is available (>= 3 valid beams) and water
track if it is not.

Also updated the plugin to add the raw_velocity and raw_velocity_covar
values to the ROS message (will use bottom track values if they are
available and water track values if they are not).