Incorporated ocean current into the DVL Gazebo plugin
The plugin subscribes to a gazebo ocean current topic (default is
hydrodynamics/current_velocity but can be specified in an
"ocean_current_topic" element in the plugin element in the SDF).
Messages published to the topic are of type gazebo/msgs/Vector3d
and specify the ocean current components in Gazebo's ENU frame.
To calculate the DVL's water tracking solution, the ocean current
is subtracted from the linear velocity of the link to which the DVL
is mounted.  The remaining velocities are then applied to each DVL
beam and the water track solution is computed in the same manner
as the bottom track solution.