Added current profiling to the DVL
Current profile is published as a dsl_sensor_msgs/Adcp message to the
DVL topic with "_current" appended.  For now the capability is implemented
entirely in the ROS plugin based on the beam-specific water speed
computed by the Gazebo plugin.  For each velocity bin, Gaussian noise
is added to each "true" beam velocity (from the Gazebo sensor) and
multiplied by the beam's unit vector to get a noisy velocity in the
beam's frame for publication in the ROS message.

No promise that the numbers are right, but if the beam velocity itself
is being computed correctly (it is done the same way that bottom-track
velocity is computed), then it should be ok.  It needs verification
from someone more familiar with the instruments themselves, though.