Updated DVL global current calculation
If a single global ocean-current is not received (e.g., the plugin
is not providing it), then the stratified current will be averaged
for the global current.  If a global current is being provided by
the plugin, it will be used instead (i.e., plugin-provided global
current will take precedence over the stratified current).

The global current is currently utilized in the velocity calculation
when in water tracking mode.  The stratified current (if available)
is always used in the current profiling calculation.